lots of changes to support libscsynth linked into SuperCollider.app
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@31 a380766d-ff14-0410-b294-a243070f3f08

+++ source/server/SC_Lib.cpp
@@ -75,14 +75,14 @@ SC_NamedObj::~SC_NamedObj()
 
 void SC_NamedObj::SetName(const int32 *inName)
 {
-	if (str4len(inName) > kSCNameLen) return;
+	if (str4len(inName) > (int)kSCNameLen) return;
 	str4cpy(mName, inName);
 	mHash = Hash(mName);
 }
 
 void SC_NamedObj::SetName(const char *inName)
 {
-	if (str4len(inName) > kSCNameLen) return;
+	if (str4len(inName) > (int)kSCNameLen) return;
 	str4cpy(mName, inName);
 	mHash = Hash(mName);
 }
